% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphenatr-package.r
\docType{package}
\name{hyphenatr}
\alias{hyphenatr}
\alias{hyphenatr-package}
\title{Tools to Hyphenate Strings Using the 'Hunspell' Hyphenation Library}
\description{
Identifying hyphenation points in strings can be useful for both
text processing and display functions. The 'Hunspell' hyphenation library
provides tools to perform hyphenation using custom language rule dictionaries.
Words can be hyphenated directly or split into hyphenated component strings
for further processing.
}
\author{
Bob Rudis (@hrbrmstr)
}

