% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hyphendict_is_loaded}
\alias{hyphendict_is_loaded}
\title{Test to see if the hyphenation language rules dictionary is loaded}
\usage{
hyphendict_is_loaded()
}
\description{
Test to see if the hyphenation language rules dictionary is loaded
}
\note{
The \code{en_US} hyphenation rules dictionary is loaded by default
}

