% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphenatr-package.r
\docType{package}
\name{hyphenatr}
\alias{hyphenatr}
\alias{hyphenatr-package}
\title{A package to HYPHENATE ALL THE THINGS}
\description{
A package to HYPHENATE ALL THE THINGS
}
\author{
Bob Rudis (@hrbrmstr)
}

