% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphen.r
\name{switch_dict}
\alias{switch_dict}
\title{Switch hyphen rules language}
\usage{
switch_dict(lang)
}
\arguments{
\item{lang}{hyphen rules language}
}
\description{
See \code{References} for the list. For valid values, use
\code{list_dicts()}.
}
\references{
\url{http://www.textcontrol.com/en_US/downloads/dictionaries/}
}

