% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyphen.r
\name{switch_dict}
\alias{switch_dict}
\title{Switch hyphen rules language}
\usage{
switch_dict(lang)
}
\arguments{
\item{lang}{hyphen rules language}
}
\description{
See \code{References} for the source of the hyphenation rules. For valid
values, use \code{list_dicts()}.
}
\note{
The \code{en_US} hyphenation rules dictionary is loaded by default
}
\examples{
switch_dict("hu_HU")
}
\references{
\url{https://github.com/LibreOffice/dictionaries}
}

